Add files via upload
In this Project, I worked with COVID19 dataset, published by John Hopkins University, which consist of the data related to cumulative number of confirmed cases, per day, in each Country. Also we have another dataset consist of various life factors, scored by the people living in each country around the globe. I merged these two datasets to see if there is any relationship between the spread of the the virus in a country and how happy people are, living in that country.